EA 23.320 Nyaightly Patch 1

+++ Elin/QuestRandom.cs
@@ -14,7 +14,7 @@ public class QuestRandom : Quest
 
 	public override void OnDropReward()
 	{
-		long num = (long)bonusMoney * (long)(55 + difficulty * 15) / 100;
+		long num = bonusMoney * (55 + difficulty * 15) / 100;
 		int num2 = MathEx.ClampToInt(num + rewardMoney, 10000000);
 		if (num2 > 0)
 		{
